requirements.txt
typing-extensions
langchain-openai
langchain-experimental
pandas
yfinance
praw
feedparser
stockstats
eodhd
langgraph>=0.5.4,<0.6.0
chromadb
setuptools
backtrader
akshare
tushare
finnhub-python
parsel
requests
tqdm
pytz
psycopg[binary]
redis
chainlit>=2.5.5; python_version < "3.12"
rich
questionary
langchain_anthropic
langchain-google-genai
jugaad-data
sentence-transformers
fastapi
uvicorn
pydantic
python-dotenv
